requirements.txt
beautifulsoup4==4.12.3
mysql-connector-python==8.3.0
Pillow==10.3.0
playwright==1.44.0
python-dotenv==1.0.1
requests==2.32.3
pytest==8.2.2
